Fix to #32570 - Test/Cleanup: some adhoc tests that are provider only because hardcoded column types etc can be moved from provider to relational/core
Also removed some tests from InMemory QBT - those tests were duplicated from SqlServer QBT but have been moved to AdHoc test suites and run for all providers.
