+++ test/EFCore.InMemory.FunctionalTests/Query/QueryBugsInMemoryTest.cs
@@ -660,231 +660,6 @@ public class QueryBugsInMemoryTest : IClassFixture<InMemoryFixture>
 
     #endregion
 
-    #region Bug19708
-
-    [ConditionalFact]
-    public virtual void GroupJoin_SelectMany_in_defining_query_is_flattened()
-    {
-        using (CreateScratch<MyContext19708>(Seed19708, "19708"))
-        {
-            using var context = new MyContext19708();
-
-            var query = context.Set<CustomerView19708>().ToList();
-
-            Assert.Collection(
-                query,
-                t => AssertCustomerView(t, 1, "First", 1, "FirstChild"),
-                t => AssertCustomerView(t, 2, "Second", 2, "SecondChild1"),
-                t => AssertCustomerView(t, 2, "Second", 3, "SecondChild2"),
-                t => AssertCustomerView(t, 3, "Third", null, ""));
-        }
-
-        static void AssertCustomerView(
-            CustomerView19708 actual,
-            int id,
-            string name,
-            int? customerMembershipId,
-            string customerMembershipName)
-        {
-            Assert.Equal(id, actual.Id);
-            Assert.Equal(name, actual.Name);
-            Assert.Equal(customerMembershipId, actual.CustomerMembershipId);
-            Assert.Equal(customerMembershipName, actual.CustomerMembershipName);
-        }
-    }
-
-    private class MyContext19708 : DbContext
-    {
-        // ReSharper disable once UnusedAutoPropertyAccessor.Local
-        public DbSet<Customer19708> Customers { get; set; }
-        // ReSharper disable once UnusedAutoPropertyAccessor.Local
-        public DbSet<CustomerMembership19708> CustomerMemberships { get; set; }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-            => optionsBuilder
-                .UseInternalServiceProvider(InMemoryFixture.DefaultServiceProvider)
-                .UseInMemoryDatabase("19708");
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-            => modelBuilder.Entity<CustomerView19708>().HasNoKey().ToInMemoryQuery(Build_Customers_Sql_View_InMemory());
-
-        private Expression<Func<IQueryable<CustomerView19708>>> Build_Customers_Sql_View_InMemory()
-        {
-            Expression<Func<IQueryable<CustomerView19708>>> query = () =>
-                from customer in Customers
-                join customerMembership in CustomerMemberships on customer.Id equals customerMembership.CustomerId into
-                    nullableCustomerMemberships
-                from customerMembership in nullableCustomerMemberships.DefaultIfEmpty()
-                select new CustomerView19708
-                {
-                    Id = customer.Id,
-                    Name = customer.Name,
-                    CustomerMembershipId = customerMembership != null ? customerMembership.Id : default(int?),
-                    CustomerMembershipName = customerMembership != null ? customerMembership.Name : ""
-                };
-            return query;
-        }
-    }
-
-    private static void Seed19708(MyContext19708 context)
-    {
-        var customer1 = new Customer19708 { Name = "First" };
-        var customer2 = new Customer19708 { Name = "Second" };
-        var customer3 = new Customer19708 { Name = "Third" };
-
-        var customerMembership1 = new CustomerMembership19708 { Name = "FirstChild", Customer = customer1 };
-        var customerMembership2 = new CustomerMembership19708 { Name = "SecondChild1", Customer = customer2 };
-        var customerMembership3 = new CustomerMembership19708 { Name = "SecondChild2", Customer = customer2 };
-
-        context.AddRange(customer1, customer2, customer3);
-        context.AddRange(customerMembership1, customerMembership2, customerMembership3);
-
-        context.SaveChanges();
-    }
-
-    private class Customer19708
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-    }
-
-    private class CustomerMembership19708
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public int CustomerId { get; set; }
-        public Customer19708 Customer { get; set; }
-    }
-
-    private class CustomerView19708
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public int? CustomerMembershipId { get; set; }
-        public string CustomerMembershipName { get; set; }
-    }
-
-    #endregion
-
-    #region Issue21768
-
-    [ConditionalFact]
-    public virtual void Using_explicit_interface_implementation_as_navigation_works()
-    {
-        using (CreateScratch<MyContext21768>(t => { }, "21768"))
-        {
-            using var context = new MyContext21768();
-            Expression<Func<IBook21768, BookViewModel21768>> projection = b => new BookViewModel21768
-            {
-                FirstPage = b.FrontCover.Illustrations.FirstOrDefault(i => i.State >= IllustrationState21768.Approved) != null
-                    ? new PageViewModel21768
-                    {
-                        Uri = b.FrontCover.Illustrations.FirstOrDefault(i => i.State >= IllustrationState21768.Approved).Uri
-                    }
-                    : null,
-            };
-
-            var result = context.Books.Where(b => b.Id == 1).Select(projection).SingleOrDefault();
-        }
-    }
-
-    private class BookViewModel21768
-    {
-        public PageViewModel21768 FirstPage { get; set; }
-    }
-
-    private class PageViewModel21768
-    {
-        public string Uri { get; set; }
-    }
-
-    private interface IBook21768
-    {
-        public int Id { get; set; }
-
-        public IBookCover21768 FrontCover { get; }
-        public int FrontCoverId { get; set; }
-
-        public IBookCover21768 BackCover { get; }
-        public int BackCoverId { get; set; }
-    }
-
-    private interface IBookCover21768
-    {
-        public int Id { get; set; }
-        public IEnumerable<ICoverIllustration21768> Illustrations { get; }
-    }
-
-    private interface ICoverIllustration21768
-    {
-        public int Id { get; set; }
-        public IBookCover21768 Cover { get; }
-        public int CoverId { get; set; }
-        public string Uri { get; set; }
-        public IllustrationState21768 State { get; set; }
-    }
-
-    private class Book21768 : IBook21768
-    {
-        public int Id { get; set; }
-
-        public BookCover21768 FrontCover { get; set; }
-        public int FrontCoverId { get; set; }
-
-        public BookCover21768 BackCover { get; set; }
-        public int BackCoverId { get; set; }
-
-        IBookCover21768 IBook21768.FrontCover
-            => FrontCover;
-
-        IBookCover21768 IBook21768.BackCover
-            => BackCover;
-    }
-
-    private class BookCover21768 : IBookCover21768
-    {
-        public int Id { get; set; }
-        public ICollection<CoverIllustration21768> Illustrations { get; set; }
-
-        IEnumerable<ICoverIllustration21768> IBookCover21768.Illustrations
-            => Illustrations;
-    }
-
-    private class CoverIllustration21768 : ICoverIllustration21768
-    {
-        public int Id { get; set; }
-        public BookCover21768 Cover { get; set; }
-        public int CoverId { get; set; }
-        public string Uri { get; set; }
-        public IllustrationState21768 State { get; set; }
-
-        IBookCover21768 ICoverIllustration21768.Cover
-            => Cover;
-    }
-
-    private enum IllustrationState21768
-    {
-        New,
-        PendingApproval,
-        Approved,
-        Printed
-    }
-
-    private class MyContext21768 : DbContext
-    {
-        // ReSharper disable once UnusedAutoPropertyAccessor.Local
-        public DbSet<Book21768> Books { get; set; }
-        public DbSet<BookCover21768> BookCovers { get; set; }
-        public DbSet<CoverIllustration21768> CoverIllustrations { get; set; }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-            => optionsBuilder
-                .UseInternalServiceProvider(InMemoryFixture.DefaultServiceProvider)
-                .UseInMemoryDatabase("21768");
-    }
-
-    #endregion
-
     #region Issue21803
 
     [ConditionalFact]
@@ -1014,218 +789,6 @@ public class QueryBugsInMemoryTest : IClassFixture<InMemoryFixture>
             => optionsBuilder
                 .UseInternalServiceProvider(InMemoryFixture.DefaultServiceProvider)
                 .UseInMemoryDatabase("20729");
-    }
-
-    #endregion
-
-    #region Issue19253
-
-    [ConditionalFact]
-    public virtual void Concat_combines_nullability_of_entity_shapers()
-    {
-        using (CreateScratch<MyContext19253>(Seed19253, "19253"))
-        {
-            using var context = new MyContext19253();
-
-            Expression<Func<A19253, string>> leftKeySelector = x => x.forkey;
-            Expression<Func<B19253, string>> rightKeySelector = y => y.forkey;
-
-            var query = context.A.GroupJoin(
-                    context.B,
-                    leftKeySelector,
-                    rightKeySelector,
-                    (left, rightg) => new { left, rightg })
-                .SelectMany(
-                    r => r.rightg.DefaultIfEmpty(),
-                    (x, y) => new JoinResult19253<A19253, B19253> { Left = x.left, Right = y })
-                .Concat(
-                    context.B.GroupJoin(
-                            context.A,
-                            rightKeySelector,
-                            leftKeySelector,
-                            (right, leftg) => new { leftg, right })
-                        .SelectMany(
-                            l => l.leftg.DefaultIfEmpty(),
-                            (x, y) => new JoinResult19253<A19253, B19253> { Left = y, Right = x.right })
-                        .Where(z => z.Left.Equals(null)))
-                .ToList();
-
-            Assert.Equal(3, query.Count);
-        }
-    }
-
-    [ConditionalFact]
-    public virtual void Union_combines_nullability_of_entity_shapers()
-    {
-        using (CreateScratch<MyContext19253>(Seed19253, "19253"))
-        {
-            using var context = new MyContext19253();
-
-            Expression<Func<A19253, string>> leftKeySelector = x => x.forkey;
-            Expression<Func<B19253, string>> rightKeySelector = y => y.forkey;
-
-            var query = context.A.GroupJoin(
-                    context.B,
-                    leftKeySelector,
-                    rightKeySelector,
-                    (left, rightg) => new { left, rightg })
-                .SelectMany(
-                    r => r.rightg.DefaultIfEmpty(),
-                    (x, y) => new JoinResult19253<A19253, B19253> { Left = x.left, Right = y })
-                .Union(
-                    context.B.GroupJoin(
-                            context.A,
-                            rightKeySelector,
-                            leftKeySelector,
-                            (right, leftg) => new { leftg, right })
-                        .SelectMany(
-                            l => l.leftg.DefaultIfEmpty(),
-                            (x, y) => new JoinResult19253<A19253, B19253> { Left = y, Right = x.right })
-                        .Where(z => z.Left.Equals(null)))
-                .ToList();
-
-            Assert.Equal(3, query.Count);
-        }
-    }
-
-    [ConditionalFact]
-    public virtual void Except_combines_nullability_of_entity_shapers()
-    {
-        using (CreateScratch<MyContext19253>(Seed19253, "19253"))
-        {
-            using var context = new MyContext19253();
-
-            Expression<Func<A19253, string>> leftKeySelector = x => x.forkey;
-            Expression<Func<B19253, string>> rightKeySelector = y => y.forkey;
-
-            var query = context.A.GroupJoin(
-                    context.B,
-                    leftKeySelector,
-                    rightKeySelector,
-                    (left, rightg) => new { left, rightg })
-                .SelectMany(
-                    r => r.rightg.DefaultIfEmpty(),
-                    (x, y) => new JoinResult19253<A19253, B19253> { Left = x.left, Right = y })
-                .Except(
-                    context.B.GroupJoin(
-                            context.A,
-                            rightKeySelector,
-                            leftKeySelector,
-                            (right, leftg) => new { leftg, right })
-                        .SelectMany(
-                            l => l.leftg.DefaultIfEmpty(),
-                            (x, y) => new JoinResult19253<A19253, B19253> { Left = y, Right = x.right }))
-                .ToList();
-
-            Assert.Single(query);
-        }
-    }
-
-    [ConditionalFact]
-    public virtual void Intersect_combines_nullability_of_entity_shapers()
-    {
-        using (CreateScratch<MyContext19253>(Seed19253, "19253"))
-        {
-            using var context = new MyContext19253();
-
-            Expression<Func<A19253, string>> leftKeySelector = x => x.forkey;
-            Expression<Func<B19253, string>> rightKeySelector = y => y.forkey;
-
-            var query = context.A.GroupJoin(
-                    context.B,
-                    leftKeySelector,
-                    rightKeySelector,
-                    (left, rightg) => new { left, rightg })
-                .SelectMany(
-                    r => r.rightg.DefaultIfEmpty(),
-                    (x, y) => new JoinResult19253<A19253, B19253> { Left = x.left, Right = y })
-                .Intersect(
-                    context.B.GroupJoin(
-                            context.A,
-                            rightKeySelector,
-                            leftKeySelector,
-                            (right, leftg) => new { leftg, right })
-                        .SelectMany(
-                            l => l.leftg.DefaultIfEmpty(),
-                            (x, y) => new JoinResult19253<A19253, B19253> { Left = y, Right = x.right }))
-                .ToList();
-
-            Assert.Single(query);
-        }
-    }
-
-    private class MyContext19253 : DbContext
-    {
-        // ReSharper disable once UnusedAutoPropertyAccessor.Local
-        public DbSet<A19253> A { get; set; }
-        // ReSharper disable once UnusedAutoPropertyAccessor.Local
-        public DbSet<B19253> B { get; set; }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-            => optionsBuilder
-                .UseInternalServiceProvider(InMemoryFixture.DefaultServiceProvider)
-                .UseInMemoryDatabase("19253");
-    }
-
-    private class JoinResult19253<TLeft, TRight>
-    {
-        public TLeft Left { get; set; }
-
-        public TRight Right { get; set; }
-    }
-
-    private class A19253
-    {
-        public int Id { get; set; }
-        public string a { get; set; }
-        public string a1 { get; set; }
-        public string forkey { get; set; }
-    }
-
-    private class B19253
-    {
-        public int Id { get; set; }
-        public string b { get; set; }
-        public string b1 { get; set; }
-        public string forkey { get; set; }
-    }
-
-    private static void Seed19253(MyContext19253 context)
-    {
-        var tmp_a = new A19253[]
-        {
-            new()
-            {
-                a = "a0",
-                a1 = "a1",
-                forkey = "a"
-            },
-            new()
-            {
-                a = "a2",
-                a1 = "a1",
-                forkey = "d"
-            },
-        };
-        var tmp_b = new B19253[]
-        {
-            new()
-            {
-                b = "b0",
-                b1 = "b1",
-                forkey = "a"
-            },
-            new()
-            {
-                b = "b2",
-                b1 = "b1",
-                forkey = "c"
-            },
-        };
-        context.A.AddRange(tmp_a);
-        context.B.AddRange(tmp_b);
-
-        context.SaveChanges();
     }
 
     #endregion
